+++ src/scripts/nori/model/ApplicationModel.js
@@ -4,8 +4,7 @@ define('nori/model/ApplicationModel',
     var _this,
       _appMapCollectionList = Object.create(null),
       _appMapList = Object.create(null),
-      _appEventConstants = require('nori/events/EventConstants'),
-      _dispatcher = require('nori/utils/Dispatcher');
+      _appEventConstants = require('nori/events/EventConstants')
 
     function initializeApplicationModel() {
       _this = this;