+++ src/scripts/nori/utils/Renderer.js
@@ -9,11 +9,10 @@ define('nori/utils/Renderer',
 
     var _appEvents         = require('nori/events/EventCreator'),
         _appEventConstants = require('nori/events/EventConstants'),
-        _dispatcher        = require('nori/utils/Dispatcher'),
         _domUtils          = require('nudoru/browser/DOMUtils');
 
     function initialize() {
-      _dispatcher.subscribe(_appEventConstants.RENDER_VIEW, render);
+      Nori.dispatcher().subscribe(_appEventConstants.RENDER_VIEW, render);
     }
 
     function render(payload) {