+++ changelog.md
@@ -6,4 +6,5 @@
     - Exposed dispatcher() from Nori
     - Nori eventCreator use Nori.dispatcher() rather than req it
     - Nori eventCreator create evtObj, dispatch it and return it
-    
+    - Event creator, added: change route, view changed to mobile / desktop
+    - refactored all to call Nori.dispatcher() rather than requiring dispatcher